python_core/requirements.txt
openai-whisper
pysubs2
python-dotenv
google-generativeai
google-adk

